fftw and fftwf enable threading support
Build also the libfftw3_threads.so
and libfftw3f_threads.so to allow using
multiple threads when running the fftw applications.

Applications wanting to take advantage for this
needs to link both to libfftw3.so and libfftw3_threads.so
(and libfftw3f.so and libfftw3f_threads.so)

For usage example see the rocFFT/clients/tests
or hipFFT/clients/tests as an example how to
enable the multithreading.

Signed-off-by: Mika Laitio <mika.laitio@amd.com>