Introduce ability for prereleases (#1720)
This adds the ability to build prereleases, pushed to a separate bucket.
Prereleases will be versioned as X.Y.ZrcYYYYMMDD whereas the version for
future nightly releases is changed to X.Y.ZaYYYYMMDD. When merging this,
the version number should be raised from 7.9 to e.g. 7.10 as new
nightlies otherwise do not take precedence over the previously build
nightlies.