Switch Python index pages to relative links (#1812)
`manage.py` creates an index.html file for every package which is placed
in a subdirectory named after the package. Packages are located one
level above whereas links where abolsute. This refactors to use relative
links instead as this allows to move the index to a different directory
without rewriting.

This is compliant with PEP 503 [1]:
URLs may be either absolute or relative as long as they point to the
correct location.

[1] https://peps.python.org/pep-0503/