Add check_path_lengths.py utility script. (#1794)
This helps spot long paths and can be customized to run as part of a
pre-commit check.

This could also be achieved with other existing CLI tools but having it
in a Python script has been useful on projects like IREE, which this
file was ported from:
https://github.com/iree-org/iree/blob/main/build_tools/scripts/check_path_lengths.py
(I authored it there also).

Sample output:
https://gist.github.com/ScottTodd/ef2b6618313d641cb30cc3e4d4c8fbaf