Modularize TheRock manifest generation into schema + writer (#1861)
## Motivation

Recognized one comment missed to address from previous PR:
https://github.com/ROCm/TheRock/pull/1790#discussion_r2441463829

These are not functional changes. It improves the code readability with
two functions in the main.

```
manifest = build_manifest_schema(repo_root, the_rock_commit)
write_manifest_json(out_path, manifest)
```

## Test Plan
Test in local and CI.
Check the CI Summary to make sure the generated manifest files exist in
the links.


## Test Result
For local, the manifest was generated
For CI, the manifest was generated which can be reach from the summary:
https://github.com/ROCm/TheRock/actions/runs/18695993535?pr=1861

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.