Fixing Release Workflow (#1964)
## Motivation

In the previous update, `linux_portable_build.py` invoked the
`linux_portable_build_in_container.sh` script using the --manylinux
argument. However, the correct approach is to enable manylinux mode by
setting the `MANYLINUX=1` environment variable. This issue arose because
the change was inadvertently mixed with another modification in
`build_configure.py`, which legitimately uses the `--manylinux`
argument. As a result, the two mechanisms were conflated, leading to
this error.

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.