Add miopen-plugin project to TheRock (#1811)
## Motivation

Follow-up after adding hipDNN to TheRock.
These changes add the MIOpen_plugin project to TheRock as well.

MIOpen_plugin depends on hipDNN & MIOpen

## Technical Details

MIOpen_plugin is a separate plugin project to add support for operations
to hipDNN.
This PR adds this project, and sets up an example for how hipDNN plugins
can be installed in the future.

## Test Plan

Tests are added as part of this PR.

## Test Result

Tests are successfully passing.

---------

Co-authored-by: Samuel Reeder <41528605+SamuelReeder@users.noreply.github.com>
Co-authored-by: Marius Brehler <marius.brehler@amd.com>