[MIOpen][CI] Filter large per-hardware MIOpen artifacts to only the requested ASIC(s) (#2039)
## Motivation

MIOpen has a number of large hardware specific artifacts including
compiled kernel databases (KDB's), Find Databases (FDB's) and Perf
Databases (Perf DB's). KDB's are stored in LFS. In TheRock, artifacts
for all hardware that MIOpen supports are included in a build regardless
of which hardware was specified for the build. Since these artifacts are
large (tens to hundres of megabytes), this wastes space and slows down
CI due to downloading artifacts that won't be used. Issue #1848 was
created for this.

## Technical Details

I added support for filtering artifacts by hardware platform to MIOpen's
build system in ROCm/rocm-libraries#2262. This PR adds support for
TheRock CI to use this feature.

## Test Plan

* Full CI run
* Inspect artifacts included in builds

## Test Result

* Full CI run pending

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.