Add debug tools feature and ROCdbgapi to TheRock (Linux)
* Motivation

Add the debug-tools feature that will serve as the home to all debug tools
projects we have: ROCdbgapi, ROCgdb and ROCR-debug-agent.

This commit adds the feature along with ROCdbgapi.

Please note this is restricted to Linux builds for now. Windows builds will
come at a later stage.

* Technical Details

This commit creates the minimal set of changes required to support the ROCm
debug tools in TheRock build platform.

It adds the following:

- New debug-tools features/subdirectory (updated main CMakeLists.txt)
- Adds ROCdbgapi as a submodule using github's amd-staging branch.
- Adds the main debug-tools/CMakeLists.txt file with instructions on how to
build the debug tools. At the moment this is just ROCdbgapi, named amd-dbgapi.
- Updates build_tools/fetch_sources.py
- Updates the documentation.
- Adds ROCdbgapi artifacts.

NOTE 1: ROCdbgapi will be added as a submodule of TheRock only temporarily until
a decision is made on the topic. It may be the case we will want to
migrate it to rocm-systems. But adding it as a submodule now allows us to
integrate the component into TheRock without having to wait for all the
rocm-systems migration steps to take place.

NOTE 2: To make integration and testing faster, we have a local
amd-dbgapi.cmake file that gets copied into the source directory before
builds. Later on these cmake changes will be integrated into ROCdbgapi
and the local file will be removed.

NOTE 3: Artifact scripting changes will come at a later date after the
testing requirements become clear.

* Test Plan

- Validate that ROCdbgapi builds and pulls its libbacktrace dependency.
- Validate that ROCdbgapi's shared object exists and matches the name.

* Test Result

All tests passed.