Remove 0004-Fix-assembler-error-in-pal-trap-handler.patch (#2041)
## Motivation

The build system carries a patch to the CRL trap handler for gfx11. This
was to work-around a bug that happened when upgrading the trap handler
from the ROCr: filtering gfx9 support caused an label to be missing.
This patch is not required any-more, so it can be dropped.

## Technical Details

The CRL fix for this made sure the required label was defined, while the
patch present in TheRock updated the branch instruction to point to
another label name (placed at the same point in the program). Those two
fixes are equivalent, so it is un-necessary to continue to carry this
patch in TheRock.

The original issue looked like:

```
  s_branch foo

[...]
.bar:
```

The fix in the CRL repo was to do:

```
  s_branch foo

[...]
.foo:
.bar:
```

while the patch maintained in TheRock did:

```
  s_barnch bar

.bar:
```

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.