Add libbacktrace to common sysdeps
ROCdbgapi links in libbacktrace to aid debugging of crashes. So add
libbacktrace to sysdeps/common.

The sources are downloaded from top-of-tree github.

We do some light smoke testing to make sure the archive gets built.

NOTE: We're downloading the packaged sources from github, but we should
setup a S3 cache to download from.