Add rocr-debug-agent to TheRock (Linux)
* Motivation

Add rocr-debug-agent to the debug tools feature, to be built alongside the
ROCdbgapi.

Please note this is restricted to Linux builds for now. Windows builds will
come at a later stage.

* Technical Details

- Add rocr-debug-agent as a submodule using github's amd-staging branch.
- Adjust debug-tools/CMakeLists.txt to handle the new rocr-debug-agent
component.
- Add rocr-debug-agent artifacts.

NOTE 1: Similarly to ROCdbgapi, rocr-debug-agent is integrated as a module only
temporarily so we can test the integration. It will be migrated to
rocm-systems later on. When that happens, we can update the build recipes.

NOTE 2: As with ROCdbgapi, we have a local cmake build file that gets copied
prior to building the component. This is to speed up integration and testing
and the local cmake file will be removed in the future and its modification
pushed to rocr-debug-agent.

NOTE 3: Artifact scripting changes will come at a later date after the
testing requirements become clear.

* Test Plan

- Validate that rocr-debug-agent builds and pulls its amd-dbgapi dependency.
- Validate that rocr-debug-agent shared object and test binary exists and
matches the required names.

* Test Result

All tests passed.