Fixes for the gpu selection logic
Selected GPU used for running tests needs to be available both in
the list of all GPUs visible and in the list of GPUs supported by
the currently tested pytorch build.

In addition the gpu index list used for setting the HIP_VISIBLE_DEVICES
needs to be based on to the index of all visible gpus.
(Instead of using the gpus supported by pytorch as a base for the index)

Signed-off-by: Mika Laitio <mika.laitio@amd.com>