Revert RCCL to previous commit at d23d18f and bump to theRock for the latest build (#2244)
## Motivation

We identify there's RCCL issue from last week's (11/11) merge at
[acdafac](https://github.com/ROCm/rccl/tree/acdafac49f8b87c8704b3447a3ed3fa0c429a34e).

So we revert the RCCL commit to previous one at
[d23d18f](https://github.com/ROCm/rccl/tree/d23d18f423d661556a5091f0dc9f27dec2adf67f).

We hope to include this reverted RCCL to the latest build and deliver it
to customer.

## Technical Details
Revert back the comm-libs to
[5fdaea5](https://github.com/ROCm/TheRock/commit/5fdaea53fbce3a5b3d78d1ef9969ef8484096a70)
and that includes previous RCCL commit d23d18f.

Please review whether there's anything missing or misconfigured. Thank
you!