[torch] Add --pytorch-dir test arg, replacing --the-rock-root-dir (#2217)
## Motivation

The script was previously assuming that pytorch was always checked out
to `TheRock/external-builds/pytorch/pytorch`. This path can be
customized, especially when testing multiple pytorch versions/branches
or when running on Windows where that default path is too long.

This is part of making it easier to reproduce pytorch test failures, see
https://github.com/ROCm/TheRock/issues/2173

## Test Plan

* Sanity check on CI:
https://github.com/ROCm/TheRock/actions/runs/19549864812/job/55978264497

* Tested by running the script locally with 

    ```bash
python
~/dev/projects/TheRock/external-builds/pytorch/run_linux_pytorch_tests.py
\
      --amdgpu-family gfx110X-dgpu \
      --pytorch-dir ~/dev/projects/pytorch \
      &> logs/tests_torch2.9_$(date +%F)_01.txt
    ```
    
(I'd love for some automated way to auto-increment that `01` each run...
hmm... maybe put the HH/MM/SS or unix timestamp too?)

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Laura Promberger <laura.promberger@amd.com>