Skip CI for changes to dockerfiles. (#2366)
This will save some CI resources when editing dockerfiles.

We could set up a different workflow that tries to build the dockerfiles
and _maybe_ runs CI workflows using the development image, but I think
we can revisit this filtering if/when we do that. Perhaps if we
centralized the image pin (outside of workflow yml files) we could have
workflows redirect that pin with a workflow step somehow?